betterapi_docker/requirements.txt
api_analytics==1.0.19
apscheduler==3.10.0
cloudscraper==1.2.68
fastapi==0.91.0
fastapi_analytics==1.0.17
fastapi_queue==0.1.1
gTTS==2.3.1
loguru==0.6.0
markdownify==0.11.6
Pillow==9.4.0
pydantic==1.10.4
PyVirtualDisplay==3.0
ratelimit==2.2.1
requests==2.27.1
selenium==4.8.0
slowapi==0.1.7
starlette==0.24.0
undetected_chromedriver==3.4.2
urllib3==1.26.9
uvicorn==0.15.0
youdotcom==1.2.1
redis>=4.2.0rc1